[#30] Feat: 모든 게시판에 Board 컴포넌트 적용
