+++ src/app/board/free/page.tsx
@@ -1,32 +1,11 @@
-import {
-  QueryClient,
-  dehydrate,
-  HydrationBoundary,
-} from '@tanstack/react-query';
-import { headers } from 'next/headers';
-import WriteButton from '@/components/Board/WriteButton';
-import { fetchPosts } from '@/service/firebase/firebase-firestore';
-import Posts from '@/components/Board/Posts';
 import BoardTitle from '@/components/Board/BoardTitle';
+import Board from '@/components/Board/Board';
 
 export default async function FreeBoardPage() {
-  const headersList = headers();
-  const path = headersList.get('x-pathname') || '';
-  const category = path.split('/')[2];
-  const queryClient = new QueryClient();
-
-  await queryClient.prefetchQuery({
-    queryKey: ['board', category],
-    queryFn: () => fetchPosts(category),
-  });
-
   return (
-    <section>
-      <HydrationBoundary state={dehydrate(queryClient)}>
-        <BoardTitle text="자유게시판" />
-        <Posts />
-      </HydrationBoundary>
-      <WriteButton />
-    </section>
+    <main>
+      <BoardTitle text="자유게시판" />
+      <Board />
+    </main>
   );
 }
