+++ src/app/board/question/page.tsx
@@ -1,3 +1,11 @@
+import Board from '@/components/Board/Board';
+import BoardTitle from '@/components/Board/BoardTitle';
+
 export default function QuestionBoardPage() {
-  return <div>질문게시판</div>;
+  return (
+    <main>
+      <BoardTitle text="질문게시판" />
+      <Board />
+    </main>
+  );
 }
